requirements.txt
google-adk>=0.1.0
google-cloud-bigquery>=3.0.0
google-cloud-aiplatform>=1.50.0
pandas>=2.0.0
streamlit>=1.35.0
plotly>=5.20.0
db-dtypes>=1.2.0
pyarrow>=14.0.0
